+++ react/components/comparisonPageRow/ProductSummaryRow.tsx
@@ -1,12 +1,17 @@
 import React, { useState, useEffect } from 'react'
 import { pathOr, isEmpty } from 'ramda'
 import { ExtensionPoint } from 'vtex.render-runtime'
-import ComparisonContext from '../../ProductComparisonContext'
+import ComparisonContext, { Dispatch } from '../../ProductComparisonContext'
 import { useCssHandles } from 'vtex.css-handles'
 import { Checkbox } from 'vtex.styleguide'
 import ComparisonProductContext from '../../ComparisonProductContext'
 import './row.css'
 import { usePixel } from 'vtex.pixel-manager'
+import { InjectedIntlProps, injectIntl, defineMessages } from 'react-intl'
+
+interface ProductSummaryRowProps extends InjectedIntlProps {
+  isShowDifferenceDefault:boolean;
+}
 
 const CSS_HANDLES = [
   'productSummaryRowContainer',
@@ -14,11 +19,28 @@ const CSS_HANDLES = [
   'showDifferencesContainer',
 ]
 
-const ProductSummaryRow = () => {
+const setShowDifferenceFirstTime = (isShowDifferenceDefault :boolean,dispatchComparison :Dispatch) =>{
+  dispatchComparison({
+    args: {
+      showDifferences: isShowDifferenceDefault,
+    },
+    type: 'SET_SHOW_DIFFERENCES',
+  })
+}
+
+const messages = defineMessages({
+  showDifferences: {
+    defaultMessage: '',
+    id: 'store/product-comparison.product-summary-row.show-differences',
+  },
+})
+
+
+const ProductSummaryRow = ({ isShowDifferenceDefault, intl }: ProductSummaryRowProps) => {
   const cssHandles = useCssHandles(CSS_HANDLES)
-
+  
+  const [isShowDifferenceByDefault, changesChecked] = useState(isShowDifferenceDefault)
   const [showDifferences, setShowDifferences] = useState(false)
-
   const {
     useProductComparisonState,
     useProductComparisonDispatch,
@@ -36,6 +58,10 @@ const ProductSummaryRow = () => {
     comparisonData
   )
   const products = pathOr([] as ProductToCompare[], ['products'], productData)
+  
+  useEffect(() =>{
+    setShowDifferenceFirstTime(isShowDifferenceByDefault,dispatchComparison)
+  }, [])
 
   const { push } = usePixel()
 
@@ -61,6 +87,7 @@ const ProductSummaryRow = () => {
   }, [comparisonData])
 
   const onSelectorChanged = (e: { target: { checked: boolean } }) => {
+    changesChecked(!isShowDifferenceByDefault)
     dispatchComparison({
       args: {
         showDifferences: e.target.checked,
@@ -68,7 +95,6 @@ const ProductSummaryRow = () => {
       type: 'SET_SHOW_DIFFERENCES',
     })
   }
-
   return isEmpty(comparisonProducts) ? (
     <div />
   ) : (
@@ -81,7 +107,7 @@ const ProductSummaryRow = () => {
             <Checkbox
               checked={showDifferences}
               id={`id-differences`}
-              label="Show only differences"
+              label={intl.formatMessage(messages.showDifferences)}
               name={`name-differences`}
               onChange={onSelectorChanged}
               value={showDifferences}
@@ -95,5 +121,17 @@ const ProductSummaryRow = () => {
     </div>
   )
 }
-
-export default ProductSummaryRow
+ProductSummaryRow.schema = {
+  title: 'editor.product-summary-row.title',
+  description: 'editor.product-summary-row.description',
+  type: 'object',
+  properties: {
+    isShowDifferenceDefault:{
+      title:'Show difference',
+      description:'',
+      default:false,
+      type:'boolean'
+    }
+  }
+}
+export default injectIntl(ProductSummaryRow)